Added Mixin of authentication
